backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
aiohttp
numpy
python-multipart